Updating assignments/java/xor/Xor.java

+++ assignments/java/xor/Xor.java
@@ -96,7 +96,7 @@ public class Xor {
     char[][] result = new char[21][1];
     char[] resultArray = new char[21*5];
     for (int i = 0; i < letterBytes.length; i++) {
-      result[i] = Character.toChars((int) letterBytes[i]);
+      result[i] = Character.toChars(letterBytes[i]);
       // System.out.println(Character.toChars(letterBytes[i]));
     }
     for (int i = 0; i < result.length; i++) {
